requirements.txt
fastapi==0.111.0
torch==2.3.1
torchvision==0.18.1
wandb==0.17.2 
python-dotenv==1.0.1